requirements.txt
numpy
opencv-python-headless
pandas
pillow
streamlit
streamlit_drawable_canvas
tensorflow
tqdm